+++ backend-api/src/modules/users/model.ts
@@ -1,4 +1,5 @@
-import { Entity, EntityItem } from 'electrodb';
+import { Entity, type EntityItem } from 'electrodb';
+import { ulid } from 'ulid';
 
 import { dcdb, tableName } from '@/backend-api/src/lib/dynamodb';
 
@@ -85,4 +86,68 @@ export const userProfiles = new Entity(
   },
 );
 
+export const userSessions = new Entity(
+  {
+    model: {
+      service: 'thatblog',
+      entity: 'users-sessions',
+      version: '1',
+    },
+    attributes: {
+      userId: {
+        type: 'string',
+        required: true,
+      },
+      sessionId: {
+        type: 'string',
+        required: true,
+        default: () => ulid(),
+      },
+      createdAt: {
+        type: 'number',
+        required: true,
+        readOnly: true,
+        default: () => Date.now(),
+      },
+      updatedAt: {
+        type: 'number',
+        readOnly: true,
+        required: true,
+        default: () => Date.now(),
+        set: () => Date.now(),
+        watch: '*',
+      },
+      archivedAt: {
+        type: 'number',
+        set: () => Date.now(),
+      },
+    },
+    indexes: {
+      /**
+       * pk: 'USERS#${userId}',
+       * sk: 'SESSION#${sessionId}',
+       */
+      byId: {
+        pk: {
+          field: 'pk',
+          composite: ['userId'],
+          template: 'USERS#${userId}',
+          casing: 'none',
+        },
+        sk: {
+          field: 'sk',
+          composite: ['sessionId'],
+          template: 'SESSION#${sessionId}',
+          casing: 'none',
+        },
+      },
+    },
+  },
+  {
+    client: dcdb,
+    table: tableName,
+  },
+);
+
 export type UserProfileItem = EntityItem<typeof userProfiles>;
+export type UserSessionItem = EntityItem<typeof userSessions>;
